Fixed mixing of size_t and uint64_t.

+++ soft_render/include/soft_render/SR_Transform.hpp
@@ -47,13 +47,13 @@ class SR_Transform
     /**
      * Array index of a parent transformation in a Scene Graph.
      */
-    uint64_t mParentId;
+    size_t mParentId;
 
   private:
     /**
      * @brief Meta-information container.
      */
-      uint64_t mFlags;
+      uint32_t mFlags;
 
     /**
      * @brief Transformation type.
